requirements.txt
appdirs==1.4.3
astroid==2.4.2
attrs==20.3.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.3
connexion==2.7.0
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-language==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
libcst==0.3.14
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
nltk==3.5
oauthlib==3.1.0
openapi-spec-validator==0.2.9
packaging==20.3
pep517==0.8.2
progress==1.5
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.6
pyrsistent==0.17.3
PySocks==1.7.1
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.14.0
toml==0.10.2
tqdm==4.52.0
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
